+++ src/ladb_opencutlist/ruby/constants.rb
@@ -1,7 +1,7 @@
 module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
-    EXTENSION_VERSION = '3.0.1-dev'.freeze
-    EXTENSION_BUILD = '202112031529'.freeze
+    EXTENSION_VERSION = '3.0.1'.freeze
+    EXTENSION_BUILD = '202112070843'.freeze
 
 end