+++ CHANGELOG.md
@@ -6,7 +6,7 @@ in 3.x, 2.x, 1.x and 0.x versions.
 
 * 3.0.1 (WIP)
 
-  * Better selection feedbacks in cutlist, cutting diagrams and labels
+  * Better selection feedbacks in cutlist, cutting diagrams, labels and reports
   * Save default _mass_unit_ and _currency_symbol_ to model even if it is not modified from default values
 
 * 3.0.0 (2021-11-14)